Refactor vertex editor mode modules
