+++ CMakeLists.txt
@@ -178,6 +178,8 @@ target_sources(
         src/game_data_editor_selection_helpers.c
         src/game_data_editor_selection_state.c
         src/game_data_editor_vertex_mode.c
+        src/game_data_editor_vertex_selection.c
+        src/game_data_editor_vertex_shared.c
         src/game_data_editor_selection_queries.c
         src/game_data_editor_debug_primitives.c
         src/game_data_editor_grid_widget.c
