+++ src/game_data_editor_vertex_mode.c
@@ -1,9 +1,9 @@
 /**
  * @file game_data_editor_vertex_mode.c
- * @brief Editor vertex-mode selection, drag, and source vertex actions.
+ * @brief Editor vertex-mode drag handling and source vertex actions.
  */
 
-#include "game_data_internal.h"
+#include "game_data_editor_vertex_internal.h"
 
 #include <float.h>
 
@@ -32,326 +32,6 @@ static slayer3d_vec3 editor_lock_offset_to_dominant_axis(slayer3d_vec3 offset)
     return slayer3d_vec3_make(0.0f, 0.0f, offset.z);
 }
 
-static void publish_editor_vertex_drag_state(slayer3d_game_data_runtime *runtime, const editor_drag_move_state *drag)
-{
-    if (runtime == NULL || runtime->scene_state == NULL)
-        return;
-
-    const bool active = drag != NULL && drag->active && drag->vertex_drag;
-    slayer3d_properties_set_bool(runtime->scene_state, "editor.vertex.drag.active", active);
-    slayer3d_properties_set_bool(runtime->scene_state, "editor.vertex.drag.moved", active ? drag->moved : false);
-    slayer3d_properties_set_bool(runtime->scene_state, "editor.vertex.drag.axis_lock_y",
-                                 active ? drag->axis_lock_y : false);
-    slayer3d_properties_set_bool(runtime->scene_state, "editor.vertex.drag.axis_lock_dominant",
-                                 active ? drag->axis_lock_dominant : false);
-    slayer3d_properties_set_int(runtime->scene_state, "editor.vertex.drag.guide_count",
-                                active ? drag->vertex_origin_count : 0);
-    slayer3d_properties_set_vec3(runtime->scene_state, "editor.vertex.drag.offset",
-                                 active ? drag->applied_offset : slayer3d_vec3_make(0.0f, 0.0f, 0.0f));
-}
-
-static void publish_editor_vertex_move_result(slayer3d_game_data_runtime *runtime, bool valid, int source_count,
-                                              int changed_count, const char *message)
-{
-    if (runtime == NULL || runtime->scene_state == NULL)
-        return;
-    slayer3d_properties_set_bool(runtime->scene_state, "editor.vertex.move.valid", valid);
-    slayer3d_properties_set_int(runtime->scene_state, "editor.vertex.move.source_count", source_count);
-    slayer3d_properties_set_int(runtime->scene_state, "editor.vertex.move.changed_count", changed_count);
-    slayer3d_properties_set_string(runtime->scene_state, "editor.vertex.move.message", message != NULL ? message : "");
-    slayer3d_properties_set_string(runtime->scene_state, "editor.tool.last_action", message != NULL ? message : "");
-}
-
-static void clear_editor_drag_move(slayer3d_game_data_runtime *runtime)
-{
-    if (runtime == NULL)
-        return;
-    SDL_zero(runtime->editor_drag_move);
-    publish_editor_vertex_drag_state(runtime, NULL);
-}
-
-static const char *editor_metadata_stable_id(const slayer3d_game_data_editor_metadata *metadata)
-{
-    return metadata != NULL && metadata->stable_id != NULL ? metadata->stable_id : "";
-}
-
-static int editor_source_index_for_selection(const brush_world_runtime *world_runtime,
-                                             const slayer3d_game_data_editor_selection *selection)
-{
-    if (world_runtime == NULL || selection == NULL || !editor_selection_is_selectable_brush(selection))
-        return -1;
-    int source_index = editor_brush_world_find_source_box_index(world_runtime, selection->element_name);
-    if (source_index >= 0)
-        return source_index;
-    return editor_brush_world_find_source_box_index(world_runtime,
-                                                    editor_metadata_stable_id(selection->element_editor));
-}
-
-static slayer3d_vec3 editor_source_vertex_meters(const brush_world_runtime *world_runtime,
-                                                 const editor_brush_source_vertex *vertex)
-{
-    const float meters_per_unit = world_runtime != NULL && world_runtime->editor_source_meters_per_unit > 0.0f
-                                      ? world_runtime->editor_source_meters_per_unit
-                                      : 0.001f;
-    return slayer3d_vec3_make((float)vertex->coord[0] * meters_per_unit, (float)vertex->coord[1] * meters_per_unit,
-                              (float)vertex->coord[2] * meters_per_unit);
-}
-
-static int editor_collect_selected_source_indices(const slayer3d_game_data_runtime *runtime,
-                                                  const brush_world_runtime *world_runtime, int fallback_source_index,
-                                                  int *out_indices, int out_capacity)
-{
-    if (runtime == NULL || world_runtime == NULL || out_indices == NULL || out_capacity <= 0)
-        return 0;
-
-    int count = 0;
-    for (int i = 0; i < runtime->editor_selected_brush_count && count < out_capacity; ++i)
-    {
-        slayer3d_game_data_editor_selection selection =
-            resolved_editor_selection(runtime, &runtime->editor_selected_brushes[i]);
-        if (selection.world_name == NULL || SDL_strcmp(selection.world_name, world_runtime->desc.name) != 0)
-            continue;
-        const int source_index = editor_source_index_for_selection(world_runtime, &selection);
-        if (source_index >= 0)
-            out_indices[count++] = source_index;
-    }
-    if (count == 0 && fallback_source_index >= 0)
-        out_indices[count++] = fallback_source_index;
-    return count;
-}
-
-static int editor_source_vertex_shared_count(const brush_world_runtime *world_runtime, const int coord[3],
-                                             const int *selected_indices, int selected_count)
-{
-    editor_brush_source_shared_vertex
-        shared[SLAYER3D_EDITOR_SELECTED_BRUSH_CAPACITY * SLAYER3D_EDITOR_SOURCE_CONVEX_VERTEX_CAPACITY];
-    int shared_count = 0;
-    if (world_runtime == NULL || coord == NULL || selected_indices == NULL || selected_count <= 0 ||
-        !editor_brush_world_find_shared_source_vertices(world_runtime, selected_indices, selected_count, shared,
-                                                        (int)SDL_arraysize(shared), &shared_count, NULL, 0))
-    {
-        return 1;
-    }
-    for (int i = 0; i < shared_count; ++i)
-    {
-        if (shared[i].coord[0] == coord[0] && shared[i].coord[1] == coord[1] && shared[i].coord[2] == coord[2])
-            return SDL_max(shared[i].reference_count, 1);
-    }
-    return 1;
-}
-
-static bool editor_source_vertex_selection_matches(const editor_source_vertex_selection *a,
-                                                   const editor_source_vertex_selection *b)
-{
-    return a != NULL && b != NULL && a->source_index == b->source_index && a->vertex_index == b->vertex_index &&
-           SDL_strcmp(a->world_name, b->world_name) == 0 && SDL_strcmp(a->brush_stable_id, b->brush_stable_id) == 0 &&
-           SDL_strcmp(a->vertex_stable_id, b->vertex_stable_id) == 0;
-}
-
-static int editor_selected_vertex_index(const slayer3d_game_data_runtime *runtime,
-                                        const editor_source_vertex_selection *selection)
-{
-    if (runtime == NULL || selection == NULL || !editor_selected_vertices_active_for_scene(runtime))
-        return -1;
-    for (int i = 0; i < runtime->editor_selected_vertex_count; ++i)
-    {
-        if (editor_source_vertex_selection_matches(&runtime->editor_selected_vertices[i], selection))
-            return i;
-    }
-    return -1;
-}
-
-static void remove_editor_selected_vertex_at(slayer3d_game_data_runtime *runtime, int index)
-{
-    if (runtime == NULL || index < 0 || index >= runtime->editor_selected_vertex_count)
-        return;
-    for (int i = index; i + 1 < runtime->editor_selected_vertex_count; ++i)
-        runtime->editor_selected_vertices[i] = runtime->editor_selected_vertices[i + 1];
-    runtime->editor_selected_vertex_count--;
-    init_editor_source_vertex_selection(&runtime->editor_selected_vertices[runtime->editor_selected_vertex_count]);
-}
-
-static bool add_editor_selected_vertex(slayer3d_game_data_runtime *runtime,
-                                       const editor_source_vertex_selection *selection)
-{
-    if (runtime == NULL || selection == NULL ||
-        runtime->editor_selected_vertex_count >= SLAYER3D_EDITOR_SELECTED_VERTEX_CAPACITY)
-    {
-        return false;
-    }
-    if (!editor_selected_vertices_active_for_scene(runtime))
-        clear_editor_selected_vertices(runtime);
-    if (editor_selected_vertex_index(runtime, selection) >= 0)
-        return true;
-    runtime->editor_selected_vertices[runtime->editor_selected_vertex_count++] = *selection;
-    runtime->editor_selected_vertex_scene = slayer3d_game_data_active_scene(runtime);
-    return true;
-}
-
-static bool editor_coord_equal(const int a[3], const int b[3])
-{
-    return a != NULL && b != NULL && a[0] == b[0] && a[1] == b[1] && a[2] == b[2];
-}
-
-static bool editor_source_vertex_selection_from_model(const brush_world_runtime *world_runtime,
-                                                      const editor_brush_source_vertex_model *model, int vertex_index,
-                                                      editor_source_vertex_selection *out_selection)
-{
-    if (world_runtime == NULL || model == NULL || out_selection == NULL || vertex_index < 0 ||
-        vertex_index >= model->vertex_count)
-    {
-        return false;
-    }
-
-    init_editor_source_vertex_selection(out_selection);
-    const editor_brush_source_vertex *vertex = &model->vertices[vertex_index];
-    const editor_brush_source_box_runtime *box = NULL;
-    if (model->brush_index >= 0 && model->brush_index < world_runtime->editor_source_box_count)
-        box = &world_runtime->editor_source_boxes[model->brush_index];
-    SDL_strlcpy(out_selection->world_name, world_runtime->desc.name != NULL ? world_runtime->desc.name : "",
-                sizeof(out_selection->world_name));
-    SDL_strlcpy(out_selection->brush_name, box != NULL && box->name != NULL ? box->name : model->brush_stable_id,
-                sizeof(out_selection->brush_name));
-    SDL_strlcpy(out_selection->brush_stable_id, model->brush_stable_id, sizeof(out_selection->brush_stable_id));
-    SDL_strlcpy(out_selection->vertex_stable_id, vertex->stable_id, sizeof(out_selection->vertex_stable_id));
-    out_selection->source_index = model->brush_index;
-    out_selection->vertex_index = vertex->vertex_index;
-    out_selection->coord[0] = vertex->coord[0];
-    out_selection->coord[1] = vertex->coord[1];
-    out_selection->coord[2] = vertex->coord[2];
-    return true;
-}
-
-static float editor_vertex_handle_pick_radius(const slayer3d_game_data_runtime *runtime)
-{
-    const float grid_size =
-        runtime != NULL && runtime->scene_state != NULL
-            ? SDL_max(slayer3d_properties_get_float(runtime->scene_state, "editor.grid.size", 1.0f), 0.001f)
-            : 1.0f;
-    return SDL_max(grid_size * 0.75f, 0.05f);
-}
-
-static bool editor_add_shared_vertex_selection_group(slayer3d_game_data_runtime *runtime,
-                                                     const editor_source_vertex_selection *selection)
-{
-    if (runtime == NULL || selection == NULL)
-        return false;
-    const brush_world_runtime *world_runtime = find_brush_world_runtime(runtime, selection->world_name);
-    if (world_runtime == NULL)
-        return false;
-
-    int selected_indices[SLAYER3D_EDITOR_SELECTED_BRUSH_CAPACITY];
-    const int selected_count = editor_collect_selected_source_indices(
-        runtime, world_runtime, selection->source_index, selected_indices, SDL_arraysize(selected_indices));
-    bool added_any = false;
-    for (int i = 0; i < selected_count; ++i)
-    {
-        editor_brush_source_vertex_model model;
-        if (!editor_brush_source_box_build_vertex_model(world_runtime, selected_indices[i], &model, NULL, 0))
-            continue;
-        for (int v = 0; v < model.vertex_count; ++v)
-        {
-            if (!editor_coord_equal(model.vertices[v].coord, selection->coord))
-                continue;
-            editor_source_vertex_selection candidate;
-            if (editor_source_vertex_selection_from_model(world_runtime, &model, v, &candidate) &&
-                add_editor_selected_vertex(runtime, &candidate))
-            {
-                added_any = true;
-            }
-        }
-    }
-    return added_any;
-}
-
-static bool editor_remove_shared_vertex_selection_group(slayer3d_game_data_runtime *runtime,
-                                                        const editor_source_vertex_selection *selection)
-{
-    if (runtime == NULL || selection == NULL || !editor_selected_vertices_active_for_scene(runtime))
-        return false;
-    bool removed_any = false;
-    for (int i = runtime->editor_selected_vertex_count - 1; i >= 0; --i)
-    {
-        const editor_source_vertex_selection *candidate = &runtime->editor_selected_vertices[i];
-        if (SDL_strcmp(candidate->world_name, selection->world_name) == 0 &&
-            editor_coord_equal(candidate->coord, selection->coord))
-        {
-            remove_editor_selected_vertex_at(runtime, i);
-            removed_any = true;
-        }
-    }
-    return removed_any;
-}
-
-void publish_editor_vertex_hover_state(slayer3d_game_data_runtime *runtime,
-                                       const slayer3d_game_data_editor_selection *hover_selection)
-{
-    if (runtime == NULL || runtime->scene_state == NULL || !editor_mode_is_vertex(runtime) ||
-        !editor_selection_is_selectable_brush(hover_selection))
-    {
-        clear_editor_vertex_hover_state(runtime);
-        return;
-    }
-
-    slayer3d_game_data_editor_selection resolved = resolved_editor_selection(runtime, hover_selection);
-    const brush_world_runtime *world_runtime = find_brush_world_runtime(runtime, resolved.world_name);
-    const int source_index = editor_source_index_for_selection(world_runtime, &resolved);
-    editor_brush_source_vertex_model model;
-    if (world_runtime == NULL || source_index < 0 ||
-        !editor_brush_source_box_build_vertex_model(world_runtime, source_index, &model, NULL, 0))
-    {
-        clear_editor_vertex_hover_state(runtime);
-        return;
-    }
-
-    const slayer3d_vec3 local_point = slayer3d_vec3_sub(resolved.point, resolved.world_position);
-    int nearest_vertex = -1;
-    float nearest_distance_sq = FLT_MAX;
-    for (int i = 0; i < model.vertex_count; ++i)
-    {
-        const slayer3d_vec3 vertex = editor_source_vertex_meters(world_runtime, &model.vertices[i]);
-        const float distance_sq = slayer3d_vec3_length_squared(slayer3d_vec3_sub(vertex, local_point));
-        if (distance_sq < nearest_distance_sq)
-        {
-            nearest_distance_sq = distance_sq;
-            nearest_vertex = i;
-        }
-    }
-    if (nearest_vertex < 0)
-    {
-        clear_editor_vertex_hover_state(runtime);
-        return;
-    }
-
-    int selected_indices[SLAYER3D_EDITOR_SELECTED_BRUSH_CAPACITY];
-    const int selected_count = editor_collect_selected_source_indices(
-        runtime, world_runtime, source_index, selected_indices, SDL_arraysize(selected_indices));
-    const editor_brush_source_vertex *vertex = &model.vertices[nearest_vertex];
-    const int shared_count =
-        editor_source_vertex_shared_count(world_runtime, vertex->coord, selected_indices, selected_count);
-    const slayer3d_vec3 coord_meters = editor_source_vertex_meters(world_runtime, vertex);
-
-    slayer3d_properties_set_bool(runtime->scene_state, "editor.vertex.hover.hit", true);
-    slayer3d_properties_set_string(runtime->scene_state, "editor.vertex.hover.brush", resolved.element_name);
-    slayer3d_properties_set_string(runtime->scene_state, "editor.vertex.hover.brush_stable_id",
-                                   editor_metadata_stable_id(resolved.element_editor));
-    slayer3d_properties_set_string(runtime->scene_state, "editor.vertex.hover.vertex", vertex->stable_id);
-    slayer3d_properties_set_int(runtime->scene_state, "editor.vertex.hover.index", vertex->vertex_index);
-    slayer3d_properties_set_int(runtime->scene_state, "editor.vertex.hover.shared_count", shared_count);
-    slayer3d_properties_set_int(runtime->scene_state, "editor.vertex.hover.x", vertex->coord[0]);
-    slayer3d_properties_set_int(runtime->scene_state, "editor.vertex.hover.y", vertex->coord[1]);
-    slayer3d_properties_set_int(runtime->scene_state, "editor.vertex.hover.z", vertex->coord[2]);
-    slayer3d_properties_set_vec3(runtime->scene_state, "editor.vertex.hover.coord", coord_meters);
-    editor_source_vertex_selection hovered;
-    if (editor_source_vertex_selection_from_model(world_runtime, &model, nearest_vertex, &hovered))
-    {
-        slayer3d_properties_set_bool(runtime->scene_state, "editor.vertex.hover.selected",
-                                     editor_selected_vertex_index(runtime, &hovered) >= 0);
-    }
-}
-
-static void editor_begin_vertex_drag(slayer3d_game_data_runtime *runtime, slayer3d_input_manager *input,
-                                     const slayer3d_game_data_editor_selection *hover_selection);
 static void publish_editor_vertex_add_result(slayer3d_game_data_runtime *runtime, yyjson_val *action, bool valid,
                                              int vertex_count, int added_count, const char *message);
 static void publish_editor_vertex_add_preview(slayer3d_game_data_runtime *runtime,
@@ -361,179 +41,6 @@ static bool editor_preview_add_vertex_to_source_coord(brush_world_runtime *world
                                                       int *out_vertex_count, char *message, size_t message_size);
 static bool editor_add_vertex_to_source_coord(slayer3d_game_data_runtime *runtime, yyjson_val *action,
                                               brush_world_runtime *world_runtime, int source_index, const int coord[3]);
-static bool editor_merge_selected_vertices_to_target(slayer3d_game_data_runtime *runtime, yyjson_val *action,
-                                                     const editor_source_vertex_selection *target);
-static int editor_source_units_from_meters(const brush_world_runtime *world_runtime, float meters);
-
-static bool editor_hover_source_vertex_selection_with_distance(
-    const slayer3d_game_data_runtime *runtime, const slayer3d_game_data_editor_selection *hover_selection,
-    editor_source_vertex_selection *out_selection, float *out_distance_sq)
-{
-    if (out_selection != NULL)
-        init_editor_source_vertex_selection(out_selection);
-    if (out_distance_sq != NULL)
-        *out_distance_sq = FLT_MAX;
-    if (runtime == NULL || out_selection == NULL || !editor_mode_is_vertex(runtime) ||
-        !editor_selection_is_selectable_brush(hover_selection))
-    {
-        return false;
-    }
-
-    slayer3d_game_data_editor_selection resolved = resolved_editor_selection(runtime, hover_selection);
-    const brush_world_runtime *world_runtime = find_brush_world_runtime(runtime, resolved.world_name);
-    const int source_index = editor_source_index_for_selection(world_runtime, &resolved);
-    editor_brush_source_vertex_model model;
-    if (world_runtime == NULL || source_index < 0 ||
-        !editor_brush_source_box_build_vertex_model(world_runtime, source_index, &model, NULL, 0))
-    {
-        return false;
-    }
-
-    const slayer3d_vec3 local_point = slayer3d_vec3_sub(resolved.point, resolved.world_position);
-    int nearest_vertex = -1;
-    float nearest_distance_sq = FLT_MAX;
-    for (int i = 0; i < model.vertex_count; ++i)
-    {
-        const slayer3d_vec3 vertex = editor_source_vertex_meters(world_runtime, &model.vertices[i]);
-        const float distance_sq = slayer3d_vec3_length_squared(slayer3d_vec3_sub(vertex, local_point));
-        if (distance_sq < nearest_distance_sq)
-        {
-            nearest_distance_sq = distance_sq;
-            nearest_vertex = i;
-        }
-    }
-    if (out_distance_sq != NULL)
-        *out_distance_sq = nearest_distance_sq;
-    return editor_source_vertex_selection_from_model(world_runtime, &model, nearest_vertex, out_selection);
-}
-
-static bool editor_source_brush_is_selected(const slayer3d_game_data_runtime *runtime,
-                                            const slayer3d_game_data_editor_selection *selection)
-{
-    if (runtime == NULL || !editor_selected_brushes_active_for_scene(runtime) ||
-        !editor_selection_is_selectable_brush(selection))
-    {
-        return false;
-    }
-
-    slayer3d_game_data_editor_selection resolved = resolved_editor_selection(runtime, selection);
-    const brush_world_runtime *world_runtime = find_brush_world_runtime(runtime, resolved.world_name);
-    const int source_index = editor_source_index_for_selection(world_runtime, &resolved);
-    if (world_runtime == NULL || source_index < 0)
-        return false;
-
-    for (int i = 0; i < runtime->editor_selected_brush_count; ++i)
-    {
-        slayer3d_game_data_editor_selection selected =
-            resolved_editor_selection(runtime, &runtime->editor_selected_brushes[i]);
-        if (selected.world_name == NULL || SDL_strcmp(selected.world_name, resolved.world_name) != 0)
-            continue;
-        if (editor_source_index_for_selection(world_runtime, &selected) == source_index)
-            return true;
-    }
-    return false;
-}
-
-static bool editor_hover_source_vertex_selection(const slayer3d_game_data_runtime *runtime,
-                                                 const slayer3d_game_data_editor_selection *hover_selection,
-                                                 editor_source_vertex_selection *out_selection)
-{
-    return editor_hover_source_vertex_selection_with_distance(runtime, hover_selection, out_selection, NULL);
-}
-
-bool editor_handle_vertex_selection(slayer3d_game_data_runtime *runtime,
-                                    const slayer3d_game_data_editor_selection *hover_selection, bool select_requested,
-                                    bool *out_consumed)
-{
-    if (out_consumed != NULL)
-        *out_consumed = false;
-    if (runtime == NULL || !editor_mode_is_vertex(runtime) || !select_requested)
-        return true;
-
-    if (editor_selection_is_selectable_brush(hover_selection))
-    {
-        const slayer3d_game_data_editor_selection resolved = resolved_editor_selection(runtime, hover_selection);
-        if (editor_selected_brush_index(runtime, &resolved) < 0 && !editor_source_brush_is_selected(runtime, &resolved))
-        {
-            if (!editor_select_mode_primary_click(runtime, &resolved))
-                return false;
-            clear_editor_selected_vertices(runtime);
-            slayer3d_properties_set_string(runtime->scene_state, "editor.tool.last_action", "brush selected");
-            if (out_consumed != NULL)
-                *out_consumed = true;
-            return true;
-        }
-    }
-
-    editor_source_vertex_selection hovered;
-    float nearest_distance_sq = FLT_MAX;
-    if (!editor_hover_source_vertex_selection_with_distance(runtime, hover_selection, &hovered, &nearest_distance_sq))
-    {
-        if ((SDL_GetModState() & SDL_KMOD_SHIFT) == 0)
-            clear_editor_selected_vertices(runtime);
-        if (out_consumed != NULL)
-            *out_consumed = true;
-        return true;
-    }
-    const float handle_radius = editor_vertex_handle_pick_radius(runtime);
-    if (nearest_distance_sq > handle_radius * handle_radius)
-    {
-        if ((SDL_GetModState() & SDL_KMOD_SHIFT) == 0)
-            clear_editor_selected_vertices(runtime);
-        if (editor_selection_is_selectable_brush(hover_selection))
-        {
-            runtime->editor_active_selection = resolved_editor_selection(runtime, hover_selection);
-            runtime->editor_selection_scene = slayer3d_game_data_active_scene(runtime);
-            slayer3d_properties_set_string(runtime->scene_state, "editor.tool.last_action", "brush selected");
-        }
-        if (out_consumed != NULL)
-            *out_consumed = true;
-        return true;
-    }
-
-    const bool quick_merge =
-        (SDL_GetModState() & (SDL_KMOD_CTRL | SDL_KMOD_ALT)) != 0 && (SDL_GetModState() & SDL_KMOD_SHIFT) == 0 &&
-        editor_selected_vertices_active_for_scene(runtime) && runtime->editor_selected_vertex_count > 0;
-    if (quick_merge && editor_selected_vertex_index(runtime, &hovered) < 0)
-    {
-        if (!editor_merge_selected_vertices_to_target(runtime, NULL, &hovered))
-            return false;
-        publish_editor_selected_vertex_count(runtime);
-        publish_editor_vertex_hover_state(runtime, hover_selection);
-        if (out_consumed != NULL)
-            *out_consumed = true;
-        return true;
-    }
-
-    const bool shift = (SDL_GetModState() & SDL_KMOD_SHIFT) != 0;
-    const bool already_selected = editor_selected_vertex_index(runtime, &hovered) >= 0;
-    if (shift && already_selected)
-    {
-        (void)editor_remove_shared_vertex_selection_group(runtime, &hovered);
-        slayer3d_properties_set_string(runtime->scene_state, "editor.tool.last_action", "vertex deselected");
-    }
-    else if (already_selected)
-    {
-        slayer3d_properties_set_string(runtime->scene_state, "editor.tool.last_action", "vertex selection drag");
-        editor_begin_vertex_drag(runtime, runtime_input(runtime), hover_selection);
-    }
-    else
-    {
-        if (!shift)
-            clear_editor_selected_vertices(runtime);
-        if (!editor_add_shared_vertex_selection_group(runtime, &hovered))
-            return false;
-        slayer3d_properties_set_string(runtime->scene_state, "editor.tool.last_action",
-                                       shift ? "vertex added to selection" : "vertex selected");
-        editor_begin_vertex_drag(runtime, runtime_input(runtime), hover_selection);
-    }
-
-    publish_editor_selected_vertex_count(runtime);
-    publish_editor_vertex_hover_state(runtime, hover_selection);
-    if (out_consumed != NULL)
-        *out_consumed = true;
-    return true;
-}
 
 static int editor_source_snap_units_from_grid(const brush_world_runtime *world_runtime, float grid_size)
 {
@@ -647,176 +154,7 @@ bool editor_handle_vertex_add_to_source(slayer3d_game_data_runtime *runtime,
     return true;
 }
 
-static bool editor_project_world_to_viewport(const slayer3d_camera3d *camera, const editor_trace_viewport_config *view,
-                                             slayer3d_vec3 point, float *out_x, float *out_y)
-{
-    if (camera == NULL || view == NULL || out_x == NULL || out_y == NULL || view->width <= 0.0f || view->height <= 0.0f)
-    {
-        return false;
-    }
-
-    slayer3d_mat4 view_matrix;
-    slayer3d_mat4 projection_matrix;
-    if (!slayer3d_camera3d_compute_matrices(camera, (int)SDL_roundf(view->width), (int)SDL_roundf(view->height), 0.01f,
-                                            1000.0f, &view_matrix, &projection_matrix))
-    {
-        return false;
-    }
-
-    const slayer3d_mat4 view_projection = slayer3d_mat4_multiply(projection_matrix, view_matrix);
-    const slayer3d_vec4 clip = slayer3d_mat4_transform_vec4(view_projection, slayer3d_vec4_from_vec3(point, 1.0f));
-    if (SDL_fabsf(clip.w) <= 0.000001f)
-        return false;
-
-    const float ndc_x = clip.x / clip.w;
-    const float ndc_y = clip.y / clip.w;
-    const float ndc_z = clip.z / clip.w;
-    if (ndc_z < -1.0f || ndc_z > 1.0f)
-        return false;
-
-    *out_x = (ndc_x + 1.0f) * 0.5f * view->width;
-    *out_y = (1.0f - ndc_y) * 0.5f * view->height;
-    return true;
-}
-
-static bool editor_lasso_contains_point(const editor_drag_move_state *drag, const editor_trace_viewport_config *view,
-                                        float screen_x, float screen_y)
-{
-    if (drag == NULL || view == NULL)
-        return false;
-    const float start_x = drag->start_mouse_x - view->x;
-    const float start_y = drag->start_mouse_y - view->y;
-    const float end_x = drag->current_mouse_x - view->x;
-    const float end_y = drag->current_mouse_y - view->y;
-    const float min_x = SDL_min(start_x, end_x);
-    const float max_x = SDL_max(start_x, end_x);
-    const float min_y = SDL_min(start_y, end_y);
-    const float max_y = SDL_max(start_y, end_y);
-    return screen_x >= min_x && screen_x <= max_x && screen_y >= min_y && screen_y <= max_y;
-}
-
-static int editor_select_vertices_in_lasso(slayer3d_game_data_runtime *runtime, yyjson_val *selection_json,
-                                           const editor_drag_move_state *drag)
-{
-    if (runtime == NULL || selection_json == NULL || drag == NULL || !editor_selected_brushes_active_for_scene(runtime))
-        return 0;
-
-    yyjson_val *trace = obj_get(selection_json, "trace");
-    editor_trace_viewport_config viewport;
-    if (!editor_trace_select_viewport_at(runtime, trace, drag->start_mouse_x, drag->start_mouse_y, &viewport))
-        return 0;
-
-    slayer3d_camera3d camera;
-    if (!slayer3d_game_data_get_camera(runtime, viewport.camera, &camera))
-        return 0;
-
-    if (!drag->lasso_additive)
-        clear_editor_selected_vertices(runtime);
-
-    for (int i = 0; i < runtime->editor_selected_brush_count; ++i)
-    {
-        slayer3d_game_data_editor_selection selection =
-            resolved_editor_selection(runtime, &runtime->editor_selected_brushes[i]);
-        const brush_world_runtime *world_runtime = find_brush_world_runtime(runtime, selection.world_name);
-        const int source_index = editor_source_index_for_selection(world_runtime, &selection);
-        editor_brush_source_vertex_model model;
-        if (world_runtime == NULL || source_index < 0 ||
-            !editor_brush_source_box_build_vertex_model(world_runtime, source_index, &model, NULL, 0))
-        {
-            continue;
-        }
-
-        for (int v = 0; v < model.vertex_count; ++v)
-        {
-            const slayer3d_vec3 local = editor_source_vertex_meters(world_runtime, &model.vertices[v]);
-            const slayer3d_vec3 world = slayer3d_vec3_add(selection.world_position, local);
-            float screen_x = 0.0f;
-            float screen_y = 0.0f;
-            if (!editor_project_world_to_viewport(&camera, &viewport, world, &screen_x, &screen_y) ||
-                !editor_lasso_contains_point(drag, &viewport, screen_x, screen_y))
-            {
-                continue;
-            }
-            editor_source_vertex_selection candidate;
-            if (editor_source_vertex_selection_from_model(world_runtime, &model, v, &candidate))
-                (void)editor_add_shared_vertex_selection_group(runtime, &candidate);
-        }
-    }
-
-    publish_editor_selected_vertex_count(runtime);
-    return runtime->editor_selected_vertex_count;
-}
-
-bool editor_handle_vertex_lasso(slayer3d_game_data_runtime *runtime, yyjson_val *selection_json,
-                                const slayer3d_game_data_editor_selection *hover_selection, bool *out_consumed)
-{
-    if (out_consumed != NULL)
-        *out_consumed = false;
-    if (runtime == NULL || runtime->scene_state == NULL)
-        return true;
-
-    editor_drag_move_state *drag = &runtime->editor_drag_move;
-    if (!editor_mode_is_vertex(runtime))
-    {
-        if (drag->active && drag->vertex_lasso)
-            clear_editor_drag_move(runtime);
-        publish_editor_vertex_lasso_state(runtime, NULL, 0);
-        return true;
-    }
-
-    slayer3d_input_manager *input = runtime_input(runtime);
-    if (input == NULL)
-        return true;
-
-    const bool left_pressed = slayer3d_input_is_mouse_button_pressed(input, SDL_BUTTON_LEFT);
-    const bool left_down = slayer3d_input_is_mouse_button_down(input, SDL_BUTTON_LEFT);
-    const bool left_released = slayer3d_input_is_mouse_button_released(input, SDL_BUTTON_LEFT);
-    float mouse_x = drag->current_mouse_x;
-    float mouse_y = drag->current_mouse_y;
-    (void)slayer3d_input_get_mouse_position(input, &mouse_x, &mouse_y);
-
-    if (drag->active && drag->vertex_lasso)
-    {
-        drag->current_mouse_x = mouse_x;
-        drag->current_mouse_y = mouse_y;
-        publish_editor_vertex_lasso_state(runtime, drag, 0);
-        if (out_consumed != NULL)
-            *out_consumed = true;
-        if (left_released || !left_down)
-        {
-            const float dx = drag->current_mouse_x - drag->start_mouse_x;
-            const float dy = drag->current_mouse_y - drag->start_mouse_y;
-            const bool meaningful_drag = dx * dx + dy * dy >= 16.0f;
-            const int selected_count =
-                meaningful_drag ? editor_select_vertices_in_lasso(runtime, selection_json, drag) : 0;
-            slayer3d_properties_set_string(runtime->scene_state, "editor.tool.last_action",
-                                           meaningful_drag ? "vertex lasso selected" : "vertex lasso canceled");
-            publish_editor_vertex_lasso_state(runtime, NULL, selected_count);
-            clear_editor_drag_move(runtime);
-        }
-        return true;
-    }
-
-    const bool selectable_hover = editor_selection_is_selectable_brush(hover_selection);
-    if (left_pressed && !selectable_hover)
-    {
-        SDL_zero(*drag);
-        drag->active = true;
-        drag->vertex_lasso = true;
-        drag->lasso_additive = (SDL_GetModState() & SDL_KMOD_SHIFT) != 0;
-        drag->scene = slayer3d_game_data_active_scene(runtime);
-        drag->start_mouse_x = mouse_x;
-        drag->start_mouse_y = mouse_y;
-        drag->current_mouse_x = mouse_x;
-        drag->current_mouse_y = mouse_y;
-        publish_editor_vertex_lasso_state(runtime, drag, 0);
-        if (out_consumed != NULL)
-            *out_consumed = true;
-    }
-    return true;
-}
-
-static int editor_source_units_from_meters(const brush_world_runtime *world_runtime, float meters)
+int editor_source_units_from_meters(const brush_world_runtime *world_runtime, float meters)
 {
     const float meters_per_unit = world_runtime != NULL && world_runtime->editor_source_meters_per_unit > 0.0f
                                       ? world_runtime->editor_source_meters_per_unit
@@ -1627,8 +965,8 @@ bool slayer3d_game_data_merge_selected_editor_vertices_to_hover(slayer3d_game_da
     return editor_merge_selected_vertices_to_target(runtime, action, &target);
 }
 
-static bool editor_merge_selected_vertices_to_target(slayer3d_game_data_runtime *runtime, yyjson_val *action,
-                                                     const editor_source_vertex_selection *target)
+bool editor_merge_selected_vertices_to_target(slayer3d_game_data_runtime *runtime, yyjson_val *action,
+                                              const editor_source_vertex_selection *target)
 {
     if (runtime == NULL || target == NULL)
     {
@@ -2242,8 +1580,8 @@ static void editor_capture_vertex_drag_origins(slayer3d_game_data_runtime *runti
     }
 }
 
-static void editor_begin_vertex_drag(slayer3d_game_data_runtime *runtime, slayer3d_input_manager *input,
-                                     const slayer3d_game_data_editor_selection *hover_selection)
+void editor_begin_vertex_drag(slayer3d_game_data_runtime *runtime, slayer3d_input_manager *input,
+                              const slayer3d_game_data_editor_selection *hover_selection)
 {
     if (runtime == NULL || input == NULL || hover_selection == NULL)
         return;
